+++ src/main/java/com/sbi/example/repository/BalanceEnquiryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BalanceEnquiryRepository extends JpaRepository<BalanceEnquiryEntity, Integer> {
+public interface BalanceEnquiryRepository extends JpaRepository<BalanceEnquiryEntity, Long> {
 }
